chore: Don't lint and test when prepublish
This unnecessarily and slows down development, for example running `npm
link` it would then run the tests and lint.

- These test may fail as I could be in an development/iteration cycle
- Running tests and linting significantly slows down these commands.

When pulling down a repository and trying to install it, when you run
`npm install` the tests and linting is ran. This can be problematic as
the eslint rules and versions are not pinned to exact versions and bumps
of them can cause `npm install` to fail. Linting the files when trying
to `npm install` just adds an unnecessary obstacle.