refactor: Disallow source document to be object
I feel like allowing the source document to be a JavaScript object was a
developer convince, but it is an anti-pattern as there will be no source
maps for these types of source documents.

This will prevent any accidental parsing of an object to the parser
causing missing source maps.

BREAKING CHANGE